refactor(templates): split 'superdesk-templates' module.
`superdesk-template` becomes:
```
.
├── controllers
│   ├── CreateTemplateController.js
│   ├── TemplateMenuController.js
│   └── index.js
├── directives
│   ├── TemplateEditorModal.js
│   ├── TemplateListDirective.js
│   ├── TemplateSelectDirective.js
│   ├── TemplatesDirective.js
│   └── index.js
├── filters
│   └── index.js
├── services
│   ├── TemplatesService.js
│   └── index.js
├── styles
│   └── templates.less
├── tests
│   └── templates.spec.js
├── views
│   ├── create-template.html
│   ├── sd-template-select.html
│   ├── settings.html
│   ├── template-editor-modal.html
│   ├── template-list.html
│   └── templates.html
├── helpers.js
└── index.js

7 directories, 21 files
```
