+++ scripts/superdesk-templates/controllers/TemplateMenuController.js
@@ -1,0 +1,16 @@
+TemplateMenuController.$inject = ['$modal'];
+export function TemplateMenuController($modal) {
+    this.create = createFromItem;
+    function createFromItem(item) {
+        $modal.open({
+            templateUrl: 'scripts/superdesk-templates/views/create-template.html',
+            controller: 'CreateTemplateController',
+            controllerAs: 'template',
+            resolve: {
+                item: function() {
+                    return item;
+                }
+            }
+        });
+    }
+}
