+++ scripts/superdesk-templates/controllers/index.js
@@ -1,0 +1,6 @@
+export { CreateTemplateController } from './CreateTemplateController';
+export { TemplateMenuController } from './TemplateMenuController';
+
+// TODO: is this needed?
+TemplatesSettingsController.$inject = ['$scope'];
+export function TemplatesSettingsController($scope) {}
