+++ scripts/superdesk-templates/directives/TemplateEditorModal.js
@@ -1,0 +1,5 @@
+export function TemplateEditorModal() {
+    return {
+        templateUrl: 'scripts/superdesk-templates/views/template-editor-modal.html'
+    };
+}
