+++ scripts/superdesk-templates/directives/index.js
@@ -1,0 +1,4 @@
+export { TemplatesDirective } from './TemplatesDirective';
+export { TemplateSelectDirective } from './TemplateSelectDirective';
+export { TemplateListDirective } from './TemplateListDirective';
+export { TemplateEditorModal } from './TemplateEditorModal';
