+++ scripts/superdesk-templates/services/index.js
@@ -1,0 +1,1 @@
+export { TemplatesService } from './TemplatesService';
