+++ scripts/superdesk-templates/tests/templates.spec.js
